// src/Models/RegistrationModels/Dto.ts
export interface MassfussionFunctionResponseDTO {
  applicationCode?: string | undefined;
  applicationCodeName?: string | undefined;
  type?: string | undefined;
  claimValue?: string | undefined;
  functionDescription?: string | undefined;
}